refactor: 설정버튼 컴포넌트로 분리
