+++ movie_project/src/components/MenuButton.js
@@ -1,0 +1,56 @@
+import * as React from "react";
+import Box from "@mui/material/Box";
+import Drawer from "@mui/material/Drawer";
+import IconButton from "@mui/material/IconButton";
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemText from "@mui/material/ListItemText";
+import MenuIcon from "@mui/icons-material/Menu";
+import { Link } from "react-router-dom";
+
+export default function TemporaryDrawer() {
+  const [state, setState] = React.useState(false);
+
+  const toggleDrawer = (event) => {
+    setState(!state);
+  };
+
+  return (
+    <div>
+      <IconButton onClick={toggleDrawer}>
+        <MenuIcon fontSize="large" color="primary" />
+      </IconButton>
+      <Drawer open={state} onClose={toggleDrawer}>
+        <Box
+          role="presentation"
+          onClick={toggleDrawer}
+          onKeyDown={toggleDrawer}
+        >
+          <List>
+            <ListItem disablePadding>
+              <ListItemButton component={Link} to="/">
+                <ListItemText primary="메뉴1" />
+              </ListItemButton>
+            </ListItem>
+            <ListItem disablePadding>
+              <ListItemButton component={Link} to="/">
+                <ListItemText primary="메뉴2" />
+              </ListItemButton>
+            </ListItem>
+            <ListItem disablePadding>
+              <ListItemButton component={Link} to="/">
+                <ListItemText primary="메뉴3" />
+              </ListItemButton>
+            </ListItem>
+            <ListItem disablePadding>
+              <ListItemButton component={Link} to="/">
+                <ListItemText primary="메뉴4" />
+              </ListItemButton>
+            </ListItem>
+          </List>
+        </Box>
+      </Drawer>
+    </div>
+  );
+}
