+++ movie_project/src/components/NavBar.js
@@ -1,14 +1,71 @@
-import React, { useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Button from "react-bootstrap/Button";
-import Container from "react-bootstrap/Container";
-import Form from "react-bootstrap/Form";
-import Nav from "react-bootstrap/Nav";
-import Navbar from "react-bootstrap/Navbar";
-import NavDropdown from "react-bootstrap/NavDropdown";
-import Offcanvas from "react-bootstrap/Offcanvas";
-import "../scroll.css";
-import { GiHamburgerMenu } from "react-icons/gi";
+import * as React from "react";
+import AppBar from "@mui/material/AppBar";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import IconButton from "@mui/material/IconButton";
+import Typography from "@mui/material/Typography";
+import MenuIcon from "@mui/icons-material/Menu";
+import Container from "@mui/material/Container";
+import Button from "@mui/material/Button";
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemText from "@mui/material/ListItemText";
+import { Link } from "react-router-dom";
+import Drawer from "@mui/material/Drawer";
+import { styled, alpha } from "@mui/material/styles";
+import InputBase from "@mui/material/InputBase";
+import SearchIcon from "@mui/icons-material/Search";
+import ArrowBackIosNewIcon from "@mui/icons-material/ArrowBackIosNew";
+import SettingsDropdown from "./SettingDropdown";
+import { createTheme } from "@mui/material/styles";
+
+const pages = ["시리즈", "영화", "NEW! 요즘 대세 콘텐츠", "내가 찜한 콘텐츠"];
+
+const Search = styled("div")(({ theme }) => ({
+  position: "relative",
+  borderRadius: theme.shape.borderRadius,
+  backgroundColor: alpha(theme.palette.common.white, 0.15),
+  "&:hover": {
+    backgroundColor: alpha(theme.palette.common.white, 0.25),
+  },
+  marginLeft: 0,
+  marginRight: 20,
+  width: "100%",
+  [theme.breakpoints.up("sm")]: {
+    marginLeft: theme.spacing(1),
+    width: "auto",
+  },
+}));
+
+const SearchIconWrapper = styled("div")(({ theme }) => ({
+  padding: theme.spacing(0, 2),
+  height: "100%",
+  position: "absolute",
+  pointerEvents: "none",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+}));
+
+const StyledInputBase = styled(InputBase)(({ theme }) => ({
+  color: "inherit",
+  "& .MuiInputBase-input": {
+    padding: theme.spacing(1, 1, 1, 0),
+    // vertical padding + font size from searchIcon
+    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+    transition: theme.transitions.create("width"),
+    width: "100%",
+    [theme.breakpoints.up("sm")]: {
+      width: "12ch",
+      "&:focus": {
+        width: "20ch",
+      },
+    },
+  },
+}));
 
 const NavBar = ({ search, setSearch }) => {
   //Link 컴포넌트를 사용하지 않고 다른 페이지로 이동해야 할때 사용
@@ -16,13 +73,12 @@ const NavBar = ({ search, setSearch }) => {
   const goBack = () => {
     navigate(-1);
   };
-  const goArticles = () => {
-    //articles 경로로 이동
+  const goLogout = () => {
+    //Login 경로로 이동(로그아웃 하면 다시 뒤로가기 안됨)
     //replace: true는 페이지를 이동할때 현재 페이지를 기록에 남기지 않는다.
-    navigate("/articles", { replace: true });
+    navigate("/", { replace: true });
   };
 
-  const navbar = useRef(null);
   const [scrollPosition, setScrollPosition] = useState(0);
   const updateScroll = () => {
     setScrollPosition(window.scrollY || document.documentElement.scrollTop);
@@ -36,114 +92,160 @@ const NavBar = ({ search, setSearch }) => {
     setSearch(e.target.value);
   };
 
-  const [show, setShow] = useState(false);
+  const [state, setState] = React.useState(false);
 
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
+  const toggleDrawer = (event) => {
+    setState(!state);
+  };
 
-  const expand = "sm";
+  const [anchorElNav, setAnchorElNav] = React.useState(null);
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+
   return (
-    <>
-      <Navbar
-        fixed="top"
-        key={expand}
-        bg="dark"
-        expand={expand}
-        className="mb-3"
-      >
-        <Container fluid>
-          <Offcanvas show={show} onHide={handleClose} placement="start">
-            <Offcanvas.Header
-              closeButton
-              aria-labelledby="offcanvasNavbarLabel-expand-left"
-            >
-              <Offcanvas.Title id={`offcanvasNavbarLabel-expand-left`}>
-                Offcanvas
-              </Offcanvas.Title>
-            </Offcanvas.Header>
-            <Offcanvas.Body>
-              <Nav className="justify-content-end flex-grow-1 pe-3">
-                <Nav.Link href="#action1">Home</Nav.Link>
-                <Nav.Link href="#action2">Link</Nav.Link>
-                <NavDropdown
-                  title="Dropdown"
-                  id={`ffcanvasNavbarLabel-expand-left`}
-                >
-                  <NavDropdown.Item href="#action3">Action</NavDropdown.Item>
-                  <NavDropdown.Item href="#action4">
-                    Another action
-                  </NavDropdown.Item>
-                  <NavDropdown.Divider />
-                  <NavDropdown.Item href="#action5">
-                    Something else here
-                  </NavDropdown.Item>
-                </NavDropdown>
-              </Nav>
-            </Offcanvas.Body>
-          </Offcanvas>
-          <Nav className="justify-content-around flex-grow-1 pe-3">
-            {GiHamburgerMenu}
-            <Navbar.Brand>
-              <Nav.Link
-                style={{
-                  fontSize: "2em",
-                  color: "red",
-                  fontWeight: "bold",
-                }}
-                className="d-none d-md-flex"
-              >
-                NETFLIX
-              </Nav.Link>
-            </Navbar.Brand>
-            <Form className="d-flex mb-3 mt-3">
-              <Form.Control
-                type="search"
-                placeholder="Search"
-                className="me-2"
-                aria-label="Search"
-                onChange={goSearch}
-              />
-              <Button variant="outline-danger">Search</Button>
-            </Form>
-          </Nav>
-
-          <Navbar.Toggle aria-controls={`offcanvasNavbar-expand-right`} />
-          <Navbar.Offcanvas
-            id={`offcanvasNavbar-expand-right`}
-            aria-labelledby={`offcanvasNavbar-expand-right`}
-            placement="end"
+    <AppBar position="fixed">
+      <Container maxWidth="xl">
+        <Toolbar disableGutters>
+          <ArrowBackIosNewIcon
+            sx={{
+              display: { md: "flex", sm: "none", xs: "none" },
+              marginRight: "10px",
+            }}
+            onClick={goBack}
+          />
+          {/* <AdbIcon sx={{ display: { xs: "none", md: "flex" }, mr: 1 }} /> */}
+          <Typography
+            variant="h6"
+            noWrap
+            component="a"
+            href="/login"
+            sx={{
+              mr: 2,
+              display: { xs: "none", md: "flex" },
+              fontFamily: "helvetica",
+              fontWeight: 1000,
+              letterSpacing: ".1rem",
+              color: "red",
+              textDecoration: "none",
+              fontSize: "1.7em",
+            }}
           >
-            <Offcanvas.Header
-              closeButton
-              aria-labelledby="offcanvasNavbarLabel-expand-right"
+            NETFLIX
+          </Typography>
+          <ArrowBackIosNewIcon
+            sx={{ display: { md: "none" } }}
+            onClick={goBack}
+          />
+          <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
+            <IconButton
+              size="large"
+              aria-label="account of current user"
+              aria-controls="menu-appbar"
+              aria-haspopup="true"
+              color="inherit"
+              onClick={toggleDrawer}
             >
-              <Offcanvas.Title
-                id={`offcanvasNavbar-expand-right`}
-                className="red"
+              <MenuIcon fontSize="large" />
+            </IconButton>
+
+            <Drawer open={state} onClose={toggleDrawer}>
+              <Box
+                role="presentation"
+                onClick={toggleDrawer}
+                onKeyDown={toggleDrawer}
               >
-                NETFLIX
-              </Offcanvas.Title>
-            </Offcanvas.Header>
-            <Offcanvas.Body>
-              <Nav.Link className="d-sm-none" href="#action2">
-                Settings
-              </Nav.Link>
-              <Nav.Link className="d-sm-none" href="#action2">
-                Link
-              </Nav.Link>
-            </Offcanvas.Body>
-          </Navbar.Offcanvas>
-        </Container>
-      </Navbar>
-    </>
+                <List>
+                  <ListItem disablePadding>
+                    <ListItemButton component={Link} to="/login">
+                      <ListItemText primary="홈" />
+                    </ListItemButton>
+                  </ListItem>
+                  <ListItem disablePadding>
+                    <ListItemButton component={Link} to="/login">
+                      <ListItemText primary="시리즈" />
+                    </ListItemButton>
+                  </ListItem>
+                  <ListItem disablePadding>
+                    <ListItemButton component={Link} to="/login">
+                      <ListItemText primary="영화" />
+                    </ListItemButton>
+                  </ListItem>
+                  <ListItem disablePadding>
+                    <ListItemButton component={Link} to="/login">
+                      <ListItemText primary="NEW! 요즘 대세 컨텐츠" />
+                    </ListItemButton>
+                  </ListItem>
+                  <ListItem disablePadding>
+                    <ListItemButton component={Link} to="/login">
+                      <ListItemText primary="내가 찜한 콘텐츠" />
+                    </ListItemButton>
+                  </ListItem>
+                  <ListItem disablePadding>
+                    <ListItemButton component={Link} to="/login">
+                      <ListItemText primary="언어별로 찾아보기" />
+                    </ListItemButton>
+                  </ListItem>
+                </List>
+              </Box>
+            </Drawer>
+          </Box>
+
+          <Typography
+            variant="h5"
+            noWrap
+            component="a"
+            href="/login"
+            sx={{
+              fontFamily: "helvetica",
+              fontWeight: 1000,
+              letterSpacing: ".1rem",
+              color: "red",
+              textDecoration: "none",
+              fontSize: "1.7em",
+              mr: 2,
+              display: { sm: "flex", md: "none", xs: "none" },
+              flexGrow: 1,
+            }}
+          >
+            NETFLIX
+          </Typography>
+          <Box sx={{ flexGrow: 1, display: { xs: "none", md: "flex" } }}>
+            {pages.map((page) => (
+              <Button
+                key={page}
+                onClick={handleCloseNavMenu}
+                sx={{ my: 2, color: "white", display: "block" }}
+              >
+                {page}
+              </Button>
+            ))}
+          </Box>
+          <Search onClick={goSearch}>
+            <SearchIconWrapper>
+              <SearchIcon />
+            </SearchIconWrapper>
+            <StyledInputBase
+              placeholder="Search…"
+              inputProps={{ "aria-label": "search" }}
+              onChange={goSearch}
+            />
+          </Search>
+          <Box sx={{ flexGrow: 0 }}>
+            <SettingsDropdown />
+          </Box>
+        </Toolbar>
+      </Container>
+    </AppBar>
   );
 };
 
 export default NavBar;
 {
-  /* searchBar input의 onChange함수에 처음 값이 바뀌었을 때 state값 변경, Link를 searchResult로 이동
+  /* searchBar input의 onChange함수에 처음 값이 바뀌었을 때 부모컴포넌트의 state값 변경, 아래에 입력한 값의 변화에 따라 searchResult 띄우기
     logout버튼 누르면 Link는 / 로 이동(처음 로그인 페이지) 아니면 Navigate로 맨 처음으로 이동
-    button 버튼은 모두 li로 바꾸고 float로 띄운다음에 각각 link연결, bootstrap으로 아이콘 삽입
+    searcbar버튼을 눌렀을 때나 엔터를 쳤을 때 searchResult Link걸어서 searchResult로 이동
     
   */
 }
