+++ movie_project/src/pages/SearchResult.js
@@ -1,7 +1,7 @@
 const SearchResult = () => {
   return (
     <div>
-      <div>SEARCH RESULT</div>
+      <div style={{ marginTop: "100px", height: "1000px" }}>SEARCH RESULT</div>
     </div>
   );
 };
