--- src/Footer.js
@@ -1,9 +1,0 @@
-const Footer = () => {
-  return (
-    <div>
-      <div>footer</div>
-    </div>
-  );
-};
-
-export default Footer;
