--- src/Login.js
@@ -1,9 +1,0 @@
-const Login = () => {
-  return (
-    <div>
-      <div>LOGIN PAGE</div>
-    </div>
-  );
-};
-
-export default Login;
