--- src/Pages/Home.js
@@ -1,9 +1,0 @@
-const Home = () => {
-  return (
-    <div>
-      <div style={{ marginTop: "100px", height: "1000px" }}>HOME</div>
-    </div>
-  );
-};
-
-export default Home;
