--- src/Pages/MyPage.js
@@ -1,9 +1,0 @@
-const MyPage = () => {
-  return (
-    <div>
-      <div>MyPage</div>
-    </div>
-  );
-};
-
-export default MyPage;
