--- src/Pages/NotFound.js
@@ -1,9 +1,0 @@
-const NotFound = () => {
-  return (
-    <div>
-      <div>NotFound</div>
-    </div>
-  );
-};
-
-export default NotFound;
