--- src/Pages/SearchResult.js
@@ -1,9 +1,0 @@
-const SearchResult = () => {
-  return (
-    <div>
-      <div style={{ marginTop: "100px", height: "1000px" }}>SEARCH RESULT</div>
-    </div>
-  );
-};
-
-export default SearchResult;
