--- src/scroll.css
@@ -1,35 +1,0 @@
-.scrolledStyle  {
-    background-color: black;
-    transition: 1s;
-    color: white;
-    position:fixed;
-    top:0;
-    width:100%;
-    font-weight:500;
-    height:80px;
-  };
-
-  .original  {
-
-    position:fixed;
-    top:0;
-    width:100%;
-    font-weight:500;
-    height:80px;
-  };
-
-  ul li{
-    margin-top:25px;
-    display:block;
-    position:relative;
-    margin-top:20px;
-  }
-ul li:nth-of-type(3){
-  position: absolute;
-  right:100px;
-  top:25px;
-}
-
-.red{
-  color:red;
-}
